feat: Extract asset from proces module
